# services/openapi.yaml
openapi: 3.0.3
info:
  title: Api Performance Testing
  description: |-
    Simple API for performance testing.
  version: 1.1.0

paths:
  /hello:
    get:
      summary: Returns simple JSON object with "Hello World!" text
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          description: Internal Server Error

  /fibonacci/{n}:
    get:
      summary: Returns n-th element of Fibonacci Sequence in a JSON object
      parameters:
        - name: n
          in: path
          description: Number of Fibonacci element to return
          required: true
          schema:
            type: integer
            minimum: 0
            example: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fibonacci'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  
  /database:
    get:
      summary: Returns random element from database
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found
        '500':
          description: Internal Server Error

components:
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
          description: The message text
          example: Hello World!
      required:
        - message
    Fibonacci:
      type: object
      properties:
        number:
          type: integer
          description: The position in the Fibonacci sequence
          example: 10
        value:
          type: integer
          description: The value of the Fibonacci number at the specified position
          example: 55
      required:
        - number
        - value